Multi-Exit, Move Cube Fixes, Movement Fixes
-Implemented Shadowshift Multi-exit, using arrow keys to go further (up
arrow) or closer (down arrow) to wall, then enter to appear at the
shadow's location. Also, escape transfers the player back into the wall
from wherever they are during the exit.
-In concert with the previous change, all dynamic and moving objects are
now 'frozen' when the player is transitioning in or out, allow the
multi-exit to work with the moving objects.
-Movement cube prefab overhauled and now is grabbed by holding down E
and can only be moved forward and backward in respect to the side the
player grabbed it from. They also are affected by gravity now.
-Added grid textures in Textures folder, and they are implemented across
various prefabs.
-Renamed and tweaked Conveyor Belts/Platforms to be Pulley Systems and
Platforms, along with their scripts, respectively.
-Hotfix to the propellor system so it works Shadowshift out
-Added a new prefab for engineers only called TransitionFollow (later
will be used by VFX artist to make visual smoke effect better)
-Tweaked player base prefab by separating out mesh from parent
transform, and moved player mesh down a few pixels to seem as if the
object is actually on the ground.
-Fixed various movement bug fixes, including movement relative to camera
and player character sliding with movement.
-Fixed various multi-exit bugs as were found.
-Folder expansion and overhaul as needed. Various new prefab and scripts
folders for organizational purposes..